Fixed bug with edit task crashing after being brought from archive
App stores numbers as Strings, Firebase converts to numbers and keeps as numbers leading to conversion issues.

+++ app/src/main/java/com/clairvoyance/clairvoyance/DataField.kt
@@ -19,7 +19,7 @@ data class DataField (
             var ndata : Any?;
             ndata = when (nType) {
                 DataType.DATE -> DateTimeConverter.stringToDate(rawdata);
-                DataType.NUMBER -> rawdata.toInt()
+                DataType.NUMBER -> rawdata
                 DataType.TEXT -> rawdata
                 DataType.CHECKBOX -> Checkbox(rawdata.toBoolean(),hs["desc"]!!)
                 else -> null
